Xcode 6 Beta 4 Fixes
Float/CGFloat, Access Modifiers, Landmarks

+++ XMCircleGestureRecognizer/ViewController.swift
@@ -11,7 +11,7 @@ import UIKit
 class ViewController: UIViewController {
     
     var feedbackLabel = UILabel(frame: CGRectZero)
-    var currentValue:Float = 0.0 {
+    var currentValue:CGFloat = 0.0 {
         didSet {
             if (currentValue > 100) {
                 currentValue = 100
@@ -48,15 +48,15 @@ class ViewController: UIViewController {
         
         if let rotation = recognizer.rotation {
             currentValue += rotation.degrees / 360 * 100
-            feedbackLabel.text = feedbackLabel.text + String(format:"Value: %.2f%%", currentValue)
+            feedbackLabel.text = feedbackLabel.text + String(format:"Value: %.2f%%", Float(currentValue))
         }
         
         if let angle = recognizer.angle {
-            feedbackLabel.text = feedbackLabel.text + "\n" + String(format:"Angle: %.2f%", angle.degrees)
+            feedbackLabel.text = feedbackLabel.text + "\n" + String(format:"Angle: %.2f%", Float(angle.degrees))
         }
         
         if let distance = recognizer.distance {
-            feedbackLabel.text = feedbackLabel.text + "\n" + String(format:"Distance: %.0f%", distance)
+            feedbackLabel.text = feedbackLabel.text + "\n" + String(format:"Distance: %.0f%", Float(distance))
         }
     }
